# perf: multiply by zerofier after fold