--- a/provers/stark/src/constraints/evaluator.rs
+++ b/provers/stark/src/constraints/evaluator.rs
@@ -201,7 +201,7 @@ impl<F: IsFFTField> ConstraintEvaluator<F> {
                 // Add each term of the transition constraints to the
                 // composition polynomial, including the zerofier, the
                 // challenge and the exemption polynomial if it is necessary.
-                let acc_transition = zerofier * evaluations_transition
+                let acc_transition = evaluations_transition
                     .iter()
                     .zip(&air.context().transition_exemptions)
                     .zip(transition_coefficients)
@@ -213,9 +213,7 @@ impl<F: IsFFTField> ConstraintEvaluator<F> {
                         } else {
                             //TODO: change how exemptions are indexed!
                             if num_exemptions == 1 {
-                                acc + beta
-                                    * eval
-                                    * &transition_exemptions_evaluations[0][i]
+                                acc + beta * eval * &transition_exemptions_evaluations[0][i]
                             } else {
                                 // This case is not used for Cairo Programs, it can be improved in the future
                                 let vector = air
@@ -231,12 +229,11 @@ impl<F: IsFFTField> ConstraintEvaluator<F> {
                                     .position(|elem_2| elem_2 == exemption)
                                     .expect("is there");
 
-                                acc + beta
-                                    * eval
-                                    * &transition_exemptions_evaluations[index][i]
+                                acc + beta * eval * &transition_exemptions_evaluations[index][i]
                             }
                         }
-                    });
+                    })
+                    * zerofier;
                 // TODO: Remove clones
 
                 acc_transition + boundary