This is mostly for purism, as the gain is marginal, but we've been repeating the same multiplication by each zerofier many times. The PR fixes that.

## Type of change

- [x] Optimization

## Checklist
- [x] This change is an Optimization
  - [x] Benchmarks added/run

## Benchmark Results
```shell
# ./bench.sh 
Benchmark 1: base_seq
  Time (mean ± σ):     122.824 s ±  0.250 s    [User: 117.360 s, System: 5.456 s]
  Range (min … max):   122.539 s … 123.185 s    10 runs
 
Benchmark 2: zero_seq
  Time (mean ± σ):     119.827 s ±  0.249 s    [User: 114.387 s, System: 5.426 s]
  Range (min … max):   119.595 s … 120.479 s    10 runs
 
Summary
  zero_seq ran
    1.03 ± 0.00 times faster than base_seq

Benchmark 1: base_par
  Time (mean ± σ):     70.212 s ±  0.094 s    [User: 175.395 s, System: 7.495 s]
  Range (min … max):   70.052 s … 70.336 s    10 runs
 
Benchmark 2: zero_par
  Time (mean ± σ):     69.243 s ±  0.110 s    [User: 171.014 s, System: 7.563 s]
  Range (min … max):   69.114 s … 69.432 s    10 runs
 
Summary
  zero_par ran
    1.01 ± 0.00 times faster than base_par
```

Proved `fibonacci(70000)` on `lambdaworks-benchmarks-x86-64-dedicated`.