@MauroToscano 2024-01-19
I'm closing it, code will change a lot